Divide numerator and denominator by their GCD at compile time
Calculating the number of cycles to delay for in delay_ms and delay_us
requires multiplying by a constant and dividing by another constant. Divide
both constants by their GCD at compile time, to allow the compiler to
better optimize the multiplication and division. This reduces the overhead in
the case where the compiler does not know the length of the delay at
compile time, or otherwise fails to inline the function.