Add convenience macros to simplify calling delay functions
Calling delay functions with non-trivial inputs requires using the
turbofish operator and braces. For example, a 5 minute delay would have
to be called as delay_sec::<{ 5 * 60 }>().

Since this syntax is a bit clunky and not very user friendly, macros
are added for each delay function such that delay_sec!(5 * 60) expands
to delay_sec::<{ 5 * 60 }>()