Allow compilation for non-AVR targets
Only compile the delay_cycles module for AVR targets. This allows the
crate to be compiled for non-AVR targets (although it won't do anything).
This could be useful, for example, if this crate or a crate that depends
on it wants to run tests that run on the host machine (the machine doing
the comiplation).